refactor: option と brand はコンポーネント自体にデータを所持するように変更
