+++ tsconfig.json
@@ -10,9 +10,10 @@
     "incremental": true,
     "esModuleInterop": true,
     "module": "esnext",
-    // "resolveJsonModule": false,
     "isolatedModules": true,
-    "jsx": "preserve"
+    "jsx": "preserve",
+    "moduleResolution": "node",
+    "resolveJsonModule": true
   },
   "include": [
     "next-env.d.ts",
